Update README.md
I deleted the "Ticketing" phrase because it was not correct. We must use JIRA at the moment and are not allowed to use another ticketing system:
"####Ticketing
We suggest using a ticketing system. Which one you use doesn’t matter, just pick one: JIRA, GitHub, etc. Postcards on the wall probably aren’t enough: Ticket information needs to be captured and stored for later. 
Tickets should refer to the project that covers the work done. Checkin comments should refer to tickets."